% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_cebrap_table}
\alias{get_cebrap_table}
\title{Get tables from Cebrap's Database}
\usage{
get_cebrap_table(conn, table)
}
\arguments{
\item{conn}{A JDBC connection object created by the \code{\link{set_connection}} function}

\item{table}{A \code{string} with the name of an existing table in Cebrap's Database}
}
\description{
\code{get_cebrap_table()} fetches tables from the Cebrap's Brazilian Legislative Database.
To use it, users must provide the name of an actual table in the database and
the function will handle the SQL queries under the hood. Is it also possible
to customize the function call to select just a sample of rows or the name
of the columns.
}
\note{
Currently, \code{get_cebrap_table()} only works with the \code{Congresso_BE} database.
}
\examples{
\dontrun{
# Request a table
conn <- set_connection(login = "login", password = "pass")
senadores <- get_cebrap_table(conn, "tbl_Sen")
}

}
