% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{list_cebrap_tables}
\alias{list_cebrap_tables}
\title{List tables from Cebrap's Database}
\usage{
list_cebrap_tables(conn)
}
\arguments{
\item{conn}{A JDBC connection object created by the \code{\link{set_connection}} function}
}
\description{
\code{list_cebrap_tables()} queries and shows all tables available for fetching in the
Cebrap's Brazilian Legislative Database.
}
\examples{
\dontrun{

# List alll tables
conn <- set_connection(login = "login", password = "pass")
list_cebrap_tables(conn)
}

}
