% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{store_credentials}
\alias{store_credentials}
\title{Store credentials as ENV VARS}
\usage{
store_credentials(login, password)
}
\arguments{
\item{login}{Database login}

\item{password}{Database password}
}
\description{
\code{store_credentials()} stores login and password information needed to connect
to Cebrap's Brazilian Legislative Database as ENV VARS. Users are encourage to
use it as way to both prevent unintentional password leakages in R scripts and
to save time spent doing manual server authentication.
}
\details{
This function need to be run just once. After that, authentication information
will be saved across R sessions and all \code{cebrapdata} functions will use these
to establish connection to Cebrap's Database.
}
